requirements.txt
openai>=0.27.0
python-dotenv>=1.0.0
datasets>=2.0.0
python>=3.8